+++ packages/api/lib/testUtils.js
@@ -2,8 +2,9 @@
 
 const fs = require('fs');
 const { randomString, randomId } = require('@cumulus/common/test-utils');
-const { sqs } = require('@cumulus/common/aws');
+const { putJsonS3Object, sqs } = require('@cumulus/common/aws');
 const { createJwtToken } = require('./token');
+const { authorizedOAuthUsersKey } = require('../app/auth');
 
 const isLocalApi = () => process.env.CUMULUS_ENV === 'local';
 
@@ -184,28 +185,6 @@ function fakeExecutionFactory(status = 'completed', type = 'fakeWorkflow') {
 }
 
 /**
- * Build a user that can be authenticated against
- *
- * @param {Object} params - params
- * @param {string} params.userName - a username
- *   Defaults to a random string
- * @param {string} params.password - a password
- *   Defaults to a random string
- * @param {integer} params.expires - an expiration time for the token
- *   Defaults to one hour from now
- * @returns {Object} - a fake user
- */
-function fakeUserFactory(params = {}) {
-  const user = {
-    userName: randomId('userName'),
-    password: randomId('password'),
-    expires: Date.now() + (60 * 60 * 1000) // Default to 1 hour
-  };
-
-  return { ...user, ...params };
-}
-
-/**
  * creates fake collection records
  *
  * @param {Object} options - properties to set on the collection
@@ -257,10 +236,10 @@ function fakeAccessTokenFactory(params = {}) {
   };
 }
 
-async function createFakeJwtAuthToken({ accessTokenModel, userModel }) {
-  const userRecord = fakeUserFactory();
-  await userModel.create(userRecord);
+const setAuthorizedOAuthUsers = (users) =>
+  putJsonS3Object(process.env.system_bucket, authorizedOAuthUsersKey(), users);
 
+async function createFakeJwtAuthToken({ accessTokenModel, username }) {
   const {
     accessToken,
     refreshToken,
@@ -268,7 +247,7 @@ async function createFakeJwtAuthToken({ accessTokenModel, userModel }) {
   } = fakeAccessTokenFactory();
   await accessTokenModel.create({ accessToken, refreshToken });
 
-  return createJwtToken({ accessToken, expirationTime, username: userRecord.userName });
+  return createJwtToken({ accessToken, expirationTime, username });
 }
 
 /**
@@ -350,10 +329,10 @@ module.exports = {
   fakeRuleFactory,
   fakeRuleFactoryV2,
   fakeFileFactory,
-  fakeUserFactory,
   fakeProviderFactory,
   getSqsQueueMessageCounts,
   getWorkflowList,
   isLocalApi,
-  testEndpoint
+  testEndpoint,
+  setAuthorizedOAuthUsers
 };