+++ packages/api/models/schemas.js
@@ -36,11 +36,16 @@ module.exports.asyncOperation = {
   title: 'AsyncOperation Object',
   description: 'Cumulus API AsyncOperation Table schema',
   type: 'object',
-  required: ['createdAt', 'id', 'status', 'updatedAt'],
+  required: ['createdAt', 'id', 'status', 'updatedAt', 'description', 'operationType'],
   additionalProperties: false,
   properties: {
     createdAt: { type: 'integer' },
     id: { type: 'string' },
+    description: { type: 'string' },
+    operationType: {
+      type: 'string',
+      enum: ['ES Index', 'Bulk Granules', 'Bulk Delete', 'Kinesis Replay']
+    },
     output: {
       description: 'The result of the operation, stored as JSON',
       type: 'string'
@@ -660,15 +665,4 @@ module.exports.execution = {
     'status',
     'createdAt'
   ]
-};
-
-module.exports.user = {
-  type: 'object',
-  properties: {
-    createdAt: { type: 'integer' },
-    expires: { type: 'integer' },
-    password: { type: 'string' },
-    updatedAt: { type: 'integer' },
-    userName: { type: 'string' }
-  }
 };