+++ packages/api/tests/endpoints/test-granules.js
@@ -15,7 +15,6 @@ const { CMR } = require('@cumulus/cmr-client');
 const {
   metadataObjectFromCMRFile
 } = require('@cumulus/cmrjs/cmr-utils');
-const { DefaultProvider } = require('@cumulus/common/key-pair-provider');
 const launchpad = require('@cumulus/common/launchpad');
 const { randomString, randomId } = require('@cumulus/common/test-utils');
 
@@ -27,7 +26,8 @@ const {
   fakeAccessTokenFactory,
   fakeCollectionFactory,
   fakeGranuleFactoryV2,
-  createFakeJwtAuthToken
+  createFakeJwtAuthToken,
+  setAuthorizedOAuthUsers
 } = require('../../lib/testUtils');
 const {
   createJwtToken
@@ -37,7 +37,6 @@ const { Search } = require('../../es/search');
 process.env.AccessTokensTable = randomId('token');
 process.env.CollectionsTable = randomId('collection');
 process.env.GranulesTable = randomId('granules');
-process.env.UsersTable = randomId('users');
 process.env.stackName = randomId('stackname');
 process.env.system_bucket = randomId('system_bucket');
 process.env.TOKEN_SECRET = randomId('secret');
@@ -64,15 +63,6 @@ async function runTestUsingBuckets(buckets, testFunction) {
   } finally {
     await Promise.all(buckets.map(aws.recursivelyDeleteS3Bucket));
   }
-}
-
-/**
- * helper for cleaning up after move files tests.
- * @param {string} publicBucket - public bucket created in setupBucketsConfig
- */
-async function teardownBuckets(publicBucket) {
-  await deleteBuckets([publicBucket, process.env.system_bucket]);
-  await createBucket(process.env.system_bucket);
 }
 
 /**
@@ -110,7 +100,6 @@ let accessTokenModel;
 let granuleModel;
 let collectionModel;
 let accessToken;
-let userModel;
 
 test.before(async (t) => {
   esIndex = randomId('esindex');
@@ -138,19 +127,25 @@ test.before(async (t) => {
   granuleModel = new models.Granule();
   await granuleModel.createTable();
 
-  // create fake Users table
-  userModel = new models.User();
-  await userModel.createTable();
+  const username = randomString();
+  await setAuthorizedOAuthUsers([username]);
 
   accessTokenModel = new models.AccessToken();
   await accessTokenModel.createTable();
 
-  accessToken = await createFakeJwtAuthToken({ accessTokenModel, userModel });
+  accessToken = await createFakeJwtAuthToken({ accessTokenModel, username });
 
   // Store the CMR password
   process.env.cmr_password_secret_name = randomString();
   await aws.secretsManager().createSecret({
     Name: process.env.cmr_password_secret_name,
+    SecretString: randomString()
+  }).promise();
+
+  // Store the Launchpad passphrase
+  process.env.launchpad_passphrase_secret_name = randomString();
+  await aws.secretsManager().createSecret({
+    Name: process.env.launchpad_passphrase_secret_name,
     SecretString: randomString()
   }).promise();
 });
@@ -181,11 +176,14 @@ test.after.always(async () => {
   await collectionModel.deleteTable();
   await granuleModel.deleteTable();
   await accessTokenModel.deleteTable();
-  await userModel.deleteTable();
   await esClient.indices.delete({ index: esIndex });
   await aws.recursivelyDeleteS3Bucket(process.env.system_bucket);
   await aws.secretsManager().deleteSecret({
     SecretId: process.env.cmr_password_secret_name,
+    ForceDeleteWithoutRecovery: true
+  }).promise();
+  await aws.secretsManager().deleteSecret({
+    SecretId: process.env.launchpad_passphrase_secret_name,
     ForceDeleteWithoutRecovery: true
   }).promise();
 });
@@ -477,11 +475,6 @@ test.serial('remove a granule from CMR with launchpad authentication', async (t)
   const launchpadStub = sinon.stub(launchpad, 'getLaunchpadToken').callsFake(() => randomString());
 
   sinon.stub(
-    DefaultProvider,
-    'decrypt'
-  ).callsFake(() => Promise.resolve('fakePassword'));
-
-  sinon.stub(
     CMR.prototype,
     'deleteGranule'
   ).callsFake(() => Promise.resolve());
@@ -511,7 +504,6 @@ test.serial('remove a granule from CMR with launchpad authentication', async (t)
   process.env.cmr_oauth_provider = 'earthdata';
   launchpadStub.restore();
   CMR.prototype.deleteGranule.restore();
-  DefaultProvider.decrypt.restore();
   cmrjs.getMetadata.restore();
 });
 
@@ -804,7 +796,7 @@ test.serial('move a file and update ECHO10 xml metadata', async (t) => {
   });
 
   CMR.prototype.ingestGranule.restore();
-  await teardownBuckets(publicBucket);
+  await aws.recursivelyDeleteS3Bucket(publicBucket);
 });
 
 test.serial('move a file and update its UMM-G JSON metadata', async (t) => {
@@ -836,7 +828,7 @@ test.serial('move a file and update its UMM-G JSON metadata', async (t) => {
     return putObject({ Bucket: file.bucket, Key: file.key, Body: ummgMetadataString });
   }));
 
-  const destinationFilepath = `${process.env.stackName}/moved_granules`;
+  const destinationFilepath = `${process.env.stackName}/moved_granules/${randomString()}`;
   const destinations = [
     {
       regex: '.*.txt$',
@@ -896,7 +888,7 @@ test.serial('move a file and update its UMM-G JSON metadata', async (t) => {
   });
 
   CMR.prototype.ingestUMMGranule.restore();
-  await teardownBuckets(publicBucket);
+  await aws.recursivelyDeleteS3Bucket(publicBucket);
 });
 
 test.serial('PUT with action move returns failure if one granule file exists', async (t) => {