+++ tf-modules/archive/bootstrap.tf
@@ -41,13 +41,6 @@ data "aws_lambda_invocation" "custom_bootstrap" {
   "ResourceProperties": {
     "ElasticSearch": {
       "host": "${var.elasticsearch_hostname}"
-    },
-    "Users": {
-      "table": "${var.dynamo_tables.users.name}",
-      "records": ${jsonencode([for x in var.users : { username : x, password : "OAuth" }])}
-    },
-    "Launchpad": {
-      "Passphrase": "${var.launchpad_passphrase}"
     }
   }
 }